Move ForwardSubsystem and add a test
